File browser Load dialog, pixel font, open wav plugin.
Clicking the load wav button will open a file dialogue that allows wav, ogg or mp3 file selection. Selecting a wav file and hitting load should result in a debug.log message with the path.
Added a pixel font, now used on the timerbubble.
Open wav plugin has been imported with sample files/scene, and required bits added to "StreamingAssets" folder. Not tried yet though. Next step would be to load whatever path was returned by the file browser and pass it to the wav parser, so it'll turn it into an audio clip. Or get "File Management" which integrates it and other file types. Later down the line.